Add phpunit.xml.dist and tests/bootstrap, and remove duplicated code
