+++ tests/Middleware/ContentTypesTest.php
@@ -30,11 +30,6 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-set_include_path(dirname(__FILE__) . '/../' . PATH_SEPARATOR . get_include_path());
-
-require_once 'Slim/Slim.php';
-require_once 'Slim/Environment.php';
-
 class ContentTypesTest extends PHPUnit_Framework_TestCase {
     public function setUp() {
         ob_start();
