+++ tests/SlimTest.php
@@ -30,23 +30,6 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-set_include_path(dirname(__FILE__) . '/../' . PATH_SEPARATOR . get_include_path());
-
-require_once 'Slim/Slim.php';
-require_once 'Slim/View.php';
-require_once 'Slim/Middleware/Flash.php';
-
-//Register non-Slim autoloader
-function customAutoLoader( $class ) {
-    $file = rtrim(dirname(__FILE__), '/') . '/' . $class . '.php';
-    if ( file_exists($file) ) {
-        require $file;
-    } else {
-        return;
-    }
-}
-spl_autoload_register('customAutoLoader');
-
 //Mock custom view
 class CustomView extends Slim_View {
     function render($template) { echo "Custom view"; }
